requirements.txt
flask
flask-cors
python-dotenv
pymongo
gunicorn
google-genai
python-dotenv